requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pdfplumber==0.9.0
sentence-transformers>=2.2.2
scikit-learn==1.3.2
openai>=1.3.0
faiss-cpu==1.7.4
numpy==1.24.3
groq>=0.4.0
huggingface_hub>=0.16.4
transformers>=4.21.0
torch>=1.9.0
